test fixes for perl 5.16 and higher
Perl 5.16 seems to have broken this test strategy:

   * grab control of a CORE function
   * exercise it
   * reassign that CORE function to something else
   * exercise again, this time getting changed behavior

The tests for Functional::Utility::throttle() need to show
that code would sleep for deterministic intervals. Both of
those t/ files accomplish this by using the strategy above.
Consequently, installing this module was broken on higher
versions of perl.

This change simply fixes the tests and bumps the version
number.

Drive-by:
  - updated license stanza in pod
  - fixed link to project's repo
  - version bump for release to cpan
  - added bin/superprove, which is lame but handy